Ensuring output tree exists before attempting to drop collection
This fix avoids the seg fault that happens otherwise when a
collection is flagged to be dropped from the output tree only (case: drop).
The collection can still be dropped already from the input tree (case: ignore).

The use case of regenerating an existing reco collection should have its own
flag "regenerate" where the input file collection is disabled and output
collection enabled, with the same name. Suggest new ticket for that.