Fix getting the bar ID in the DQM code
The chosen solution is to, instead of requiring that all code downstream of hit reco
invokes the det ID wrapper, just set a bar and module id in the reco step. So two
new setters and getters for that added to TrigScintHit. And then DQM code and clustering,
tracking, etc down the line can simply use getBarID().