Update env script to deal with locally built containers
Basically, a third argument that can be used to pull from the right repo:
dev, or pro, or 'local' in which case nothing is pulled and a local
container can be used (as specified in arg 2).

In addition, some tweaks to the help command to cover the most common attempts.