Add in electron counter and multielectron capable trigger
Add two general use processors:
1. ElectronCounter which can use a simulated number of electrons or
(for now) a TS track collection to count the umber of electrons on target.
2. MultiElectronTriggerProcessor which uses the electron count set by the
ElectronCounter (and/or possibly the overlay producer, later on) to look up
which ecal energy sum threshold to use in the trigger decision. Really a minimal
change to TriggerProcessor, and would run the same, but thresolds have to be
configured as a list. Adding as a new class is minimally intrusive but we can
overwrite TriggerProcessor with this new implementation if everyone agrees.