Add in capability to run on SimTrackerHits
This is achieved by simply listing calo and tracker hits separately in the config,
and looping over the lists in turn for each event. The treatment of SimTrackerHits is
identical to that of non-Ecal SimCalorimeterHits, so it's a little repetitive. Leaving a
templated version of this producer for future work.