Make pileup start event number random
For now, this is sampled from a number up to 500. It could be the
full number of events in the file but
a) i don't know how to access that right now and
b) it doesn't seem like we need that level of randomness for a file
that will be reused 10, max 100 times. A range a factor 5 larger
than that seems like plenty.

I also tweaked the internal bunch number interval variable names
(they were sooo long). The configurable parameter names in the
python config are still long and highly descriptive.