decide on data structure for raw
raw data structure will be a std::map whose key is a std::string
identifying the value as originating from a certain type of chip (so
that it can be matched to the correct translator). The value is a buffer
holding all of the data (including a header) for that data stream for an
event.