move nextEvent call from end of loop to beginning of loop
This is just a simple off-by-one error. In production mode, we call
nextEvent after the event is processed in order to fill the objects that
were just created, but in analysis mode (which is being used here in the
overlay producer), we need to load the objects from the input file into
memory _before_ the event is processed. You can see this occur in
Framework/Process as well.