Add in a dummy 12-channel mapping too
The first column is the electronics id. This is just a simple index
representing the order the words will be written out in the simple
test beam data format.

The second column is the detector ID (bar number in simulation).

In this mapping, the detector ID follows our standard zig-zag
pattern between front and back layer, from top to bottom,
while the electronics ID goes from top to bottom of first
front, then back layer.