WIP: NOT COMPILING: drafting a new workflow
this new workflow for following the coordinates is hopefully easier to
understand. I am labeling the coordinates with respect to the center of
each layer flower as (p,q) and leaving the global coordinates as (x,y).

This will hopefully lead to easier maintenance in the future since this
class is so delicate.