use LDMX_DOCKER_TAG env var
This is just to unify how the image is selected so users can set which dev image should be used in the env of the workflow. Right now we just use ldmx/dev:latest everywhere but I could forsee us pinning to a specific dev image version in the future.