remove exception from getRunHeader and update pointer instead
This method allows the run header held within Process::run to simply be
a pointer to the object held in memory by the EventFile object. This
will enable users to update the run header in beforeNewRun even during
input-file runs.

reinsert getRunHeader so Simulator can use it

Simulator needs access to the RunHeader at the end of a run in order to
update it with event counts, this calling situation _should_ fail if the
run does not exist so the getRunHeader that fails when the run doesn't
exist is helpful to have still