Add event display making macro
This is the basic structure for making event display plots like the ones approved for testbeam data in 2022. (Getting the actual LDMX label right still requires a style helper function I haven't shared, but could, if there is demand for it. The line adding the label is commented for now.)