move closing of event files into destructor
this means the file will be closed even during exceptions and so we can
notify ROOT to do its cleanup procedures _before_ the unloading of
libraries happens.

In order to accomplish this, we needed to move the writing of the run
headers into its own function so that it can only be called during
intentional (no-exception) closing of the output file.