rework num-tries storage
We cannot follow the incrementation from within the EventHeader since it
is `Clear`ed on each event (successful or not) so we return to using the
numTries mechanic currently on trunk. Besides that, we also carry the
numTries counter across the num events completed boundary by using
modulo instead of greater-than when comparing to the maximum number of
tries. This incures that the sum of the number of tries is always equal
to the configures maxEvents.