get process manager from particle singleton
In my adventures with Geant4 developing the G4DarkBreM package, I found that recent versions (at least all since our tag of 10.2) have a method for accessing a particle's process manager from the particle itself. Since the particle is a singleton created when requested, we can always access it and ask for the process manager. If we call the function too early in processing (perhaps if we were to manually call `GammaPhysics::ConstructProcess` at the wrong time during initialization), then the manager would not be created and attached yet, but this will work if we allow G4 to call `ConstructProcess` at the right time.