add option to filter by run number as well as event number
This is helpful for people like me who have massive samples of
heavily-filtered simulations. The best way to generate these samples is
across many runs but then merge the resulting files. This results in
several runs per file and often an event from one run is aborted by the
simulation while the same event number from a different run is the event
I wish to resimulate.