force fully-specified sub-branch-names in performance tree
As pointed out by Jim Pivarski[0], we can easily modify our branch naming
scheme and then reading the branches happens without any special uproot
parsing required. This is breaking change in terms of the schema of the
TTree but since this is separate from real data and no-one is doing it,
I am putting it in now.

[0] https://github.com/scikit-hep/uproot5/discussions/1057#discussioncomment-7793747