remove generator that is not compiling and unused
we /could/ rework this generator to use the new EventFile syntax (requiring deletion to close rather than
a specific close method), but this generator has gone unused for so long I don't think it would run even
with this patch. Besides, the updated full re-simulation infrastructure allows users to effectively re-sim
from the ECal _perfectly_ so this running mode is already covered.