explicitly check if the config is readable before continuing
Previously, loading a config as a python module reported this failure
to exist check for us rather nicely and so we didn't need to worry about
doing it ourselves. Now, we need to do it ourselves since the
Py_RunAnyFile* functions do not check for existence manually and will
fail strongly if they are given a nullptr.