add a script to print markdown list of submodule states
This script mainly uses awk to look through the submodule states and
retrieve their versions. If a submodule is not on a clean version tag,
then we fall back to just linking to the commit its on.

Besides awk, this script also requires a version of git that has the
`-C` flag (>= 1.8.5, very easy on any newer systems, I've only run into issues
with this requirement on CentOS7 systems.) and all submodules to have
their remote named `origin` which is the `git` default so I expect only
super-users to see issues.