need to enter the output event file before creating and writing
ROOT requires us to be in the correct directory when we create
the output TTree for us to have it written into the correct
location.

In order to allow downstream processors to `getHistoDirectory()`
and enter that directory for the histograms we need to enter
the output event file again here so that our run tree ends up
in the correct location.