remove all CMake and Python reference to PYTHON_PACKAGE_NAME
it is fixed to LDMX and it should not be changed for protability sake.

In the process of removing this, I've also made the opinionated choice
to remove the `.in` suffix from python files that are configured by
CMake. CMake actually configures all python files regardless of their
extension while building and so we can remove that suffix and allow for
nicer syntax highlighting in our editors.

Also remove some unmainted examples from the config modules.
(Specifically Framework.examples and SimCore.examples)