add stand-a-lone processor construction ability for system installs
If the ldmx-sw headers and libraries are available in a system location
(like they are within the production images), we can pretty easily
compile a single-file processor into a library of its own from within
the python configuration.

Updating this to handle non-system installs might be difficult. I have
not checked.