remove dead code in Simulator
Simulator::onFileClose was not actually updating the run header that was written to the output file. Instead it was just updating a local copy (since there was no `&` next to the `auto`). This dead code has also been superseded by storing the number of tries from within the Framework `numTries_` and counting the number of events in the output events TTree when analyzing.