add v12 geometry with vertical TS bars included
This geometry has a slightly thicker TS pad and thus the tracker positions may have had to change ever so slightly. Doesn't matter much as this is just to share the vertical bar arrangement and the choices made for horizontal bar gaps.

I suggest trying to make this run as is and verifying that all changes are understood with the unmodified `v12`, before porting it to a version based on `v14`.