remove now non-existent acts component from linking
The ActsPluginIdentification does not exist within ACTS v36 and so
asking for it to be built just gets ignored since its not an option ACTS
looks for; however, asking to link to it errors since this component is
not a valid library that can be linked to (it is never built).

We can remove reference to this plugin and point out that the
default-build of acts appears to satisfy our needs.