draft including ratio subplot within differ when comparing
The denominator for the ratio is whatever histogram is first in the list. 

We try to inherit the color from the histogram in the ratio plot so that the legend applies.

If we want to include error bars, we should probably switch to using `hist` directly since they have better error bar handling and have ironed out a lot of the common bugs.