check for corrupted zombie files in ldmx sw (#1443)
This resolves #1406 

On trunk, the current behavior is to try to continue even after a known
corrupted file is found. This leads to wacky results. The terminal
output is very long and contains a lot of detail about errors including the
branches that are struggling to be copied into the output file. 

Even more confusingly, some events end up in the final output file but
the run does not. These events could contain data that is corrupted? I am
not sure and have no means of checking.

This commit updates the Framework to end processing if a corrupted
(zombie or no LDMX_Events TTree) file is found unless the special
`skipCorruptedInputFiles` configuration flag is set to True by the user.

We can add `p.skipCorruptedInputFiles = True` to `merge-cfg.py` and then
we continue processing after skipping the input file.
I am _intentionally_ not adding `skipCorruptedInputFiles` to the python
interface class since I believe this is an advanced option (like `testingMode`)
that should only be used by folks who know what they are doing.