Add ASAN/UBSAN recipe to just, make a CI that runs it + adding GDB (#1474)
* Add ASAN/UBSAN recipe to just, make a CI that runs it
* Add GDB options to justfile
Module operandi I propose: we dont save the log for now, we dont run the leak sanitizer part, then we run this every time a new PR comes in, and if any of those introduce a problem, this test will fail. It will not show what the issue is (given the ASAN output saving problem), but it will show that there is an issue, so we can check out the branch locally and run this and we'll see the problem.