patch current and add new shellcheck recipe
new shellcheck recipe is focused on shellcheck'ing the recipes within
the justfile itself. As they get more complicated, it is helpful to make
sure that we maintain good shell style.

Don't leave shellchecks early on an error so that we can make sure that
the temporary files that are spawned are removed.