use denv v1.1's quiet no-overwrite init flags (#1492)
* use denv v1.1.0's quiet no-overwrite init flags

this simplifies just init for newer denv and avoids a confusing error
message that is usually just passed by

* patch up new init recipe

typo was preventing getting the minor version causing that check to
always fail. also putting quotes around all the environment variables
now as well.

* have shell leave early if error

* patch current and add new shellcheck recipe

new shellcheck recipe is focused on shellcheck'ing the recipes within
the justfile itself. As they get more complicated, it is helpful to make
sure that we maintain good shell style.

Don't leave shellchecks early on an error so that we can make sure that
the temporary files that are spawned are removed.

* use xargs instead of cat

allows possibility of white-space characters in filepaths