update default LHE ID for A' to be 1023 (#1525)
* update default LHE ID for A' to be 1023

This update then makes the default configuration of the dark brem model (from our point of view) align with the libraries that are produced by v5.0 of dark-brem-lib-gen.

Users can still use earlier versions of dark-brem-lib-gen, but then they will need to change their configuration script to look for the old LHE ID 622 rather than the new one.
```python
mySim.dark_brem.model.aprime_lhe_id = 622
```

This change *does not* alter the PDG ID of the dark photon within our simulation. It will still have the PDG ID of 622 according to Geant4 and the output SimParticle.

* update to use latest release of dbgen with MG5

now that the default A' LHE ID is set to the one used by the MG5 model, updating the CI tests to use the MG5 model should work.