unify when RunHeader is printed to info logging (#1637)
* unify when RunHeader is printed to info logging

Instead of having the RunHeader be printed at different times depending on if we are in production mode (no input files) or reco mode (with input files), I move the printing of the RunHeader to the end of the `newRun` function. This will keep the printing in the same "time" for reco mode but will move it to the beginning of the run (before the event printouts) for production mode.

* Apply clang-format

* Add missing ;

* Apply clang-format

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Tamas Vami <tamas.almos.vami@cern.ch>