remove warning on checking for container
we have been well established for only using the container for building
ldmx-sw for several years now. Additionally, this check is difficult to
maintain because there isn't a standardized way to check whether you are
in a container. The CI and I get this warning when using podman because
podman does not create either of the files listed in this check.