patches to ldmx-sw following compiler update (#1640)
* mandatory patches for configure-quick build to finish

* Apply clang-format

* dont need to inherit from std::iterator anymore in C++20

* rename into config subdirectory

* WIP: middle of Python refactor avoiding meaningless class

- copy over refactor from LDMX-Softwar/fire and merge in our improvements
- build into submodule
- still need to test that it works after refactor and update to the
  newer Python C API

* Apply clang-format

* propagate Python refactor to tests

all tests compiling, Framework test passing the run

* more precise deduction of dev image major version, patch python init

remove deprecated functions and use PyConfig style configuration before
the initialization of the python interpreter

* Apply clang-format

* a bit more cleanup, don't need cmd anymore

* patch typo in config clear statement

* don't need special define anymore

* cleanup run command

- only open config file once
- close python interpreter if an exception is thrown after it is opened

* patch up new PyConfig style of configuratoin

we do not want python to parse the argv for us, we are already parsing
it and so when python parses it, the first argument is dropped as if it
was the 'python' command calling a script

* Apply clang-format

* create object to avoid accessing an optimized-out move

this is only because we erroneously have a list of parents for sim
particles when all sim particles only ever have exactly one parent

* remove warning on checking for container

we have been well established for only using the container for building
ldmx-sw for several years now. Additionally, this check is difficult to
maintain because there isn't a standardized way to check whether you are
in a container. The CI and I get this warning when using podman because
podman does not create either of the files listed in this check.

* remove getDummy and dummy constructor

it is not used anywhere in the tests

* fix typo in copying to dereference confirmed non-null pointer

* patch up dangling references in logger

- don't try to hold a reference, just use safe_extract in place when needed or copy
- define an empty object for logRules so that it is alive long enough to be used

* flag a local variable as maybe unused

older compilers didn't catch this because the code _does_ increment nPos
within the loop over adc values, but it never accesses its value for
something else

* misplaced parenthesis caught since fabs(bool) doesnt make sense

* remove meaningless check, overlayEvent_ will never be NULL

check for overlay event below when calling `nextEvent` on the overlay

* trim dead code referencing particleIndex, use range-based for

* Apply clang-format

* more careful pointer handling in KaonPhysics::DumpDeayTable

* check to make sure primaries are not NULL

* Apply clang-format

* safely dereferencing the vector branches for the compiler

* check length of vector before blinding indexing

* Apply clang-format

* use back and front as suggested by Tamas

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>