Prep for ldmx/dev v5 with some more compiler patches (#1671)
* avoid potential mem-use-after-free bug using const reference

the reference-counted touchableHandle may destruct and cause a memory
use after free when it is deallocated. We can avoid this deallocation by
using a const reference rather than a new copy.

* lots more pointer checks to show compiler that pointers are non-null

this prevents a null-pointer dereference warning in the G4String copy
constructor

* just have logging be apart of UserAction base class

* fixup some potential nullptr deref and add static

the static keyword will (hopefully) improve performance by only
retrieving the pointer when the variable is first constructed

* use `std::abs` when getting abs of a float

clang complains about this when extra warnings are enabled

* missed a potential nullptr deref warning

* Apply clang-format

* forgot another potential nullptr dereference

I feel like this is a product of how old of a Geant4 we are using
compared to how new of a compiler we are using.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>