Update HcalSimpleDigiAndRecProducer.cxx
In the energy weighted position average, the cumulated energy was used instead of the individual hit energy. This resulted in the odd distribution of RecHits. The time weighted average is affected as well.

Further, I have implemented the algorithm for the side Hcal bars but I am happy to receive feedback on how to make the code more efficient or on which parts could be left out.