Add incode comments HcalSimpleDigiAndRecProducer.cxx
Here, I added incode comments for better readability. 

Further, it might be a good idea to raise an exception if none of the known bar orientations is found. However, I'm not sure if this is necessary or even how to "artificially" trigger this exception as all hits usually have an associated bar orientation.