Update HcalSimpleDigiAndRecProducer.cxx (#1682)
* Update HcalSimpleDigiAndRecProducer.cxx

In the energy weighted position average, the cumulated energy was used instead of the individual hit energy. This resulted in the odd distribution of RecHits. The time weighted average is affected as well.

Further, I have implemented the algorithm for the side Hcal bars but I am happy to receive feedback on how to make the code more efficient or on which parts could be left out.

* Apply clang-format

* Add incode comments HcalSimpleDigiAndRecProducer.cxx

Here, I added incode comments for better readability. 

Further, it might be a good idea to raise an exception if none of the known bar orientations is found. However, I'm not sure if this is necessary or even how to "artificially" trigger this exception as all hits usually have an associated bar orientation.

* Apply clang-format

* Update HcalSimpleDigiAndRecProducer.cxx

Now a warning message is included in case no proper bar orientation could be found.

* Apply clang-format

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>