compiling libraries as *dependencies* of SimCore functions
This only works for Clang and not for GCC.
For Clang, I _think_ this works because it uniquely defines a single source of
truth of the static variable for the compiler.
For GCC, we get an "undefined symbol error" when attempting to load the
PhotoNuclearModels library. The symbol is the implementation of the
*base* class which resides in SimCore and not SimCore_PhotoNuclearModels
so I'm going to try moving the base class into the specific library now.