move Factory singleton instance into Prototype class
This allows there to be exactly one TU where the `static` variable is
defined. Before, when the `static Factory` was in the header with the
templated class itself, I believe it was being erroneously _defined_ in
multiple TUs. Now that it is only ever defined in one TU, the dynamic
factory appears to be working with both GCC and Clang with LTO.