drop one level of macro indirection
The newer FACTORY_REGISTRATION macro does not need to separate the
namespace from the class name in order to function, so I am dropping the
*_NS macros from Framework and dropping the NS argument from the
UserAction macro.

This was simple enough syntax to rewrite with sed which made it easy
enough for me to do.

  sed '/^DECLARE_/ s/_NS//; /^DECLARE_/ s/,\s*/::/'

Only applying this to files where I knew a declaration took place.