update Factory to return [[nodiscard]] std::optional
This helps us make sure that we are checking that the factory
successfully created the object we asked for.
Additionally, I separated out the warehouse_ into its own derived
Factory (FactoryWithWarehouse) that can be used simply by using a
different macro. This is nice since many of the prototypes that use the
factory already have a central storage location for the objects and
don't need the warehouse carrying around copies of the pointers itself.