Factory updates and unification (#1704)
* move library dyn loading into Process where it is called

now Framework/PluginFactory and SimCore/Factory have the same
responsibilities and so we should (hopefully) be able to replace one
with the other

* add enable_exports property

* dont compile unrelated stuff

just tring to make testing faster

* change justfile to be able to have parallel builds

don't plan to merge this, just helping myself for testing

* add status messages and a static counter inside the class

* try manually loading the SDs library

* compiling libraries as *dependencies* of SimCore functions

This only works for Clang and not for GCC.
For Clang, I _think_ this works because it uniquely defines a single source of
truth of the static variable for the compiler.
For GCC, we get an "undefined symbol error" when attempting to load the
PhotoNuclearModels library. The symbol is the implementation of the
*base* class which resides in SimCore and not SimCore_PhotoNuclearModels
so I'm going to try moving the base class into the specific library now.

* move PhotoNuclearModel base class into PhotoNuclearModels library

* move SensitiveDetector base class into SDs library

* move PrimaryGenerator base class into Generators library

* remove previous manual loading

* move XsecBiasingOperator base class into BiasOperators library

* re-enable all parts of ldmx-sw for further testing

* add ecal_pn test for Biasing UserActions, clang compilation

* move Factory singleton instance into Prototype class

This allows there to be exactly one TU where the `static` variable is
defined. Before, when the `static Factory` was in the header with the
templated class itself, I believe it was being erroneously _defined_ in
multiple TUs. Now that it is only ever defined in one TU, the dynamic
factory appears to be working with both GCC and Clang with LTO.

* Revert "add enable_exports property"

This reverts commit 03cec417430dc4e1fe6f0ad3a400b9ee17ed100b.

This additional property is not necessary and I think removing it helps
make the CMake code a bit more understandable. Trying to keep things
clean.

* mv SimCore/Factory to Framework/Factory

* SHOULD REVERT: shortening test configs since I'm just testing loading

* replace PluginFactory with templated Factory

* move registration macro definition to Factory level

remove ParserFactory now that it is redundant, replace with framework::Factory

* better exclude rules for doxygen

* rename macros to align with our coding rules/habits

* re-enable testing

* clang-format after all these changes

* update links and docs of Factory

* revert developing changes to test configs and justfile

* drop one level of macro indirection

The newer FACTORY_REGISTRATION macro does not need to separate the
namespace from the class name in order to function, so I am dropping the
*_NS macros from Framework and dropping the NS argument from the
UserAction macro.

This was simple enough syntax to rewrite with sed which made it easy
enough for me to do.

  sed '/^DECLARE_/ s/_NS//; /^DECLARE_/ s/,\s*/::/'

Only applying this to files where I knew a declaration took place.

* should be able to run tests with clang-lto build now

* Apply clang-format

* add more detailed things to check to UnableToCreate error

* add C++20 standard flag to from-file compilation command

* add extra space for readability

* update documentation comment on Analyzer::beforeNewRun

* cleanup SimCore/CMakeLists.txt since I'm already here

isolate SimCore::Geo and remove unnecessary registered_targets link

* update Factory to return [[nodiscard]] std::optional

This helps us make sure that we are checking that the factory
successfully created the object we asked for.
Additionally, I separated out the warehouse_ into its own derived
Factory (FactoryWithWarehouse) that can be used simply by using a
different macro. This is nice since many of the prototypes that use the
factory already have a central storage location for the objects and
don't need the warehouse carrying around copies of the pointers itself.

* Apply clang-format

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>