Add MIP Reconstruction in trigger (#1764)
This adds MIP reconstruction in the ECAL and HCAL and adds relevant information into the trigger ntuples. 

For ECAL MIPs, we record the length of the track, the number of hits in the track, the number of holes (length - hits), and the sum of energy in a region around the track (isolation energy). The goal with ECAL MIP tracking is to reconstruct isolated tracks with few holes in the track. To achieve this, we have two handles isolation_e_cut and hole_fraction. First, isolation_e_cut ensures that the sum of energy in a radius (currently 50mm) around a hit in the MIP track does not exceed some value (currently 180 MeV). Second, hole_fraction ensures that there are not greater than some number of holes (currently 20%) in the track. Both of these handles ensure that reconstruction keeps a good amount of the signal (MIPs) while not triggering too much on background (electrons). The background events that do pass reconstruction typically leave short (<15 layers) tracks, whereas muons leave longer tracks (>20 layers). 

For HCAL MIPs, we record the length of the track, the number of hits in the track, the number of holes (length - hits). The goal here is to reconstruct straight MIP tracks in the back HCAL. There are NOT isolation_e_cut and hole_fraction handles here but the algorithm performs well for signal while only reconstructing very few, very short track background events. 

This also adds the sum of energy in the side HCAL to the trigger ntuples.

* add trig Hcal hits

* add first pass of mip track reco

* add mip class

* add basic mip analyzer

* begin ecal mip tracking

* add ecal mip via dR and energy req

* allow seeds from any layer, loosen dR

* correct mip tracks such that only max length is recorded

* account for layer shift, min/max energy, usedhits

* begin adding side hcal sums

* add v15 to v14.detectors.valid

* side hcal sums correctly now

* minor changes with spacing

* add isolation requirment for tracks

* beginning to only consider isolated mip tracks

* move from energy prefilter to in-line filter

* add hole fraction, ntuple objects

* add hcal mips to ntuple

* cleanup debug comments

* add growing search window if holes exist

* adjust isolation energy to capture more di-muon events

* ran just format-cpp

* add pass name to exist checks and correct DECLARE_PRODUCER

* add tvami suggestions, mostly following code guidelines better

* add additional tvami suggestions, reformat TrigMipReco files

* remove comments in trigmipreco...

* add time logging and add new producers to GENIE CI

* add tomeichlersmith suggestions

* remove leftover comments

---------

Co-authored-by: Christian Herwig <t.herwig@cern.ch>